requirements.txt
asgiref==3.5.2
Django==4.0.5
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
phonenumbers==8.12.50
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
